+++ src/main/java/de/johni0702/minecraft/bobby/FakeChunkManager.java
@@ -5,6 +5,7 @@ import de.johni0702.minecraft.bobby.ext.ClientChunkManagerExt;
 import de.johni0702.minecraft.bobby.ext.LightingProviderExt;
 import de.johni0702.minecraft.bobby.mixin.BiomeAccessAccessor;
 import de.johni0702.minecraft.bobby.mixin.ClientWorldAccessor;
+import de.johni0702.minecraft.bobby.util.FileSystemUtils;
 import io.netty.util.concurrent.DefaultThreadFactory;
 import it.unimi.dsi.fastutil.longs.Long2LongMap;
 import it.unimi.dsi.fastutil.longs.Long2LongOpenHashMap;
@@ -34,6 +35,8 @@ import net.minecraft.world.chunk.light.LightingProvider;
 import net.minecraft.world.level.storage.LevelStorage;
 import org.apache.commons.lang3.tuple.Pair;
 
+import java.nio.file.Files;
+import java.nio.file.InvalidPathException;
 import java.nio.file.Path;
 import java.util.ArrayDeque;
 import java.util.ArrayList;
@@ -90,13 +93,19 @@ public class FakeChunkManager {
 
         BobbyConfig config = Bobby.getInstance().getConfig();
 
+        String serverName = getCurrentWorldOrServerName(((ClientWorldAccessor) world).getNetworkHandler());
         long seedHash = ((BiomeAccessAccessor) world.getBiomeAccess()).getSeed();
         RegistryKey<World> worldKey = world.getRegistryKey();
         Identifier worldId = worldKey.getValue();
         Path storagePath = client.runDirectory
                 .toPath()
-                .resolve(".bobby")
-                .resolve(getCurrentWorldOrServerName(((ClientWorldAccessor) world).getNetworkHandler()))
+                .resolve(".bobby");
+        if (oldFolderExists(storagePath, serverName)) {
+            storagePath = storagePath.resolve(serverName);
+        } else {
+            storagePath = FileSystemUtils.resolveSafeDirectoryName(storagePath, serverName);
+        }
+        storagePath = storagePath
                 .resolve(seedHash + "")
                 .resolve(worldId.getNamespace())
                 .resolve(worldId.getPath());
@@ -123,6 +132,14 @@ public class FakeChunkManager {
             } catch (Exception e) {
                 e.printStackTrace();
             }
+        }
+    }
+
+    private static boolean oldFolderExists(Path bobbyFolder, String name) {
+        try {
+            return Files.exists(bobbyFolder.resolve(name));
+        } catch (InvalidPathException e) {
+            return false;
         }
     }