+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 5.2.0-SNAPSHOT
 - Update to Minecraft 1.20.6
+- Fix game crashing/freezing when world name contains illegal characters (#67)
 
 ### 5.1.0
 - Add support for servers with multiple identically-named worlds (e.g. proxies) (#72)